management commands, code restructure
Added management commands to invoke the scanning and editing of backup roots.

Also changed the gbackup.py script to manage.py to keep with that django
convention, and also to make tab completion easier (doesn't conflict with the
gbackup directory)

Moved the settings to an explicit settings.py file. This makes the system play
a bit nicer with the PyCharm Django test runner, which wants an explicit
settings module. I also wanted to do this anyways for code organization reasons.

No longer implicitly migrates the database on startup. TODO: check if there are
unapplied migrations and display an error to the user.

Added routine to recursively mark all parents of invalidated FSEntry objects as
invalid.

Added Snapshot database model, unused currently.