Object.load_payload method, tests
A new method Object.load_payload sets the payload on the object and is also a
central place to do any other processing of payload data for storage in the
local database, such as computing index fields on the Object model.

This commit adds new tests for the FSEntry.scan() method.