Revert "added custom ThreadPoolExecutor to cleanup DB connections"
This reverts commit 1ac4f70589363f018f8c737ce2d7b6b0a755d965.

After some manual testing and digging around on the web and Python bug reports,
it seems that Python will automatically delete a thread's thread-local
variables from all threading.local() objects when the thread dies.

This should clean up the database connection, which lives in a thread local
variable (django.db.connections._connections) but the database connection seems
to be involved in a reference loop, and so isn't garbage collected until some
time later when the cyclic garbage collector runs. At that time, the sqlite
connection is cleaned up.

Right now the backup process spawns a thread pool. I was thinking of adding a
call to `gc.collect()` after the backup process, but I think I'll leave that
out; I don't believe this will be a problem in practice. If hundreds of threads
are being spawned and ended in a short time interval, then we could run up the
process's memory or exhaust the file descriptors, but I expect backups to be
run infrequently enough for the cyclic garbage collector to run.