fixed logic error in push_object()
If there was a IntegrityError saving the object, that means another thread
saved the object. If this happened, then the losing thread wouldn't set the
database alias on the object, and later Django would error when assigning it to
the FSEntry.obj field because the databases don't match.

This refreshes the database on the object in this case, which should be
rare-ish so the extra database read probably won't be that bad. Still, maybe
it's worth returning just the object ID instead of the full Object instance.