the restore routine no longer requires database access
There was no reason to do a DB query since all metadata is read from the
downloaded objects. This will be important later on when the local cache may be
incomplete due to a partial cache rebuild, so restores are still possible. The
local cache is really only to find the initial object that the user wants to
restore.